+++ neoforge/src/main/java/cc/cassian/pyrite/neoforge/PyriteNeoForge.java
@@ -4,6 +4,7 @@ package cc.cassian.pyrite.neoforge;
 import cc.cassian.pyrite.Pyrite;
 import cc.cassian.pyrite.registry.neoforge.BlockCreatorImpl;
 import cc.cassian.pyrite.neoforge.client.PyriteNeoForgeClient;
+import cc.cassian.pyrite.registry.neoforge.PyriteItemGroupsImpl;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.ModContainer;
@@ -23,6 +24,7 @@ public final class PyriteNeoForge {
         BlockCreatorImpl.register(eventBus);
         eventBus.addListener(BlockCreatorImpl::addSignsToSupports);
         eventBus.addListener(BlockCreatorImpl::commonSetup);
+        eventBus.addListener(PyriteItemGroupsImpl::buildContents);
 
     }
 }